requirements.txt
certifi==2023.7.22
ansible==8.4.0
ansible-core==2.15.4
certifi==2023.5.7
cffi==1.16.0
charset-normalizer==3.1.0
cryptography==41.0.4
dnspython==2.3.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
pycparser==2.21
pymongo==4.4.0
python-decouple==3.8
PyYAML==6.0.1
requests==2.31.0
urllib3==2.0.6
